Move encoder implicits to package
